+++ app/src/main/java/com/example/botb/InputManager.java
@@ -5,6 +5,7 @@ import com.example.botb.controller.GameController;
 import com.example.botb.model.Action;
 import com.example.botb.model.Board;
 import com.example.botb.model.Game;
+import com.example.botb.model.Winner;
 import com.example.botb.model.placeable.Nexus;
 import com.example.botb.model.placeable.Shield;
 import java.io.ByteArrayInputStream;
