+++ app/src/main/java/com/example/botb/model/Winner.java
@@ -1,4 +1,4 @@
-package com.example.botb;
+package com.example.botb.model;
 
 public enum Winner {
     LOCAL_PLAYER, REMOTE_PLAYER, NONE
